fix: トラック数の上限 (32 本) を撤廃し、数に比例する固定長の器を全廃する
33 本目以降のトラックが無音 (プラグインも呼ばれない) だった。トラック数と、それに比例する
ランチャー行 (512) / 内蔵 GR メーター (256) / 変調ソース (64) / プラグイン計測 (512) を
起動時の固定長の器に入れ、溢れた分を黙って捨てていたのが根 (docs/plan_unbounded_tracks.md)。

- RT の器は off-thread で曲から数えて確保し song と同じ便で届ける (scratch は追加分だけの
  成長便、ランチャーの器、変調の RT 器、FollowerMaps)。TrackScratch の 1 秒遅延線 prealloc を廃止
- GUI への telemetry は世代名の伸びる面 (TelemetryPlane) へ移し id で引く。面の id は一式
  publish し終えてから公開する。プラグイン計測はロード時に枠を割り当てる伸びる面
- solo の透過判定 (流れ込む track / 祖先 group) を compile 時の表にし、RT で Song を歩かない
- 本数に上限が無くなった毎 buffer / install の突き合わせを線形にする (ランチャー行を Song の
  並びに揃えてカーソルで組にする、adopt_state_from は前回一致位置から探す)
- worker pool 経路へ変調の刻み面を丸ごと渡す (索引 / 動く深さ / buffer 先頭位置が pool 経路で
  だけ落ち、直列と音が違っていた)
- 送り手が便の順序を守る (park 中は新しい便を先に ring へ入れない)

+++ daw_gui/src/state/transport.rs
@@ -158,10 +158,10 @@ pub struct TransportState {
     /// ~30Hz の `ModScalarsTick` ごとに差し替わり、compose 経路が
     /// `ModPlane::scalar(id)` で引く。
     pub mod_plane: common::mod_plane::ModPlane,
-    /// r.md #117: engine が publish した track ごとの鳴っているボイス `(track index, voice)`。
+    /// r.md #117: engine が publish した track ごとの鳴っているボイス `(track id, voice)`。
     /// ~30Hz の `TrackVoicesTick` ごとに差し替わり、 変調ラックが `Note` 起点ソースの
     /// カーソルをボイスごとに描くのに使う。
-    pub track_voices: Vec<(usize, common::audio_bridge::VoiceSnapshot)>,
+    pub track_voices: Vec<(u32, common::audio_bridge::VoiceSnapshot)>,
     /// `start_transport` が読み込み待ち (`pending_plugin_loads` / asset decode) で queue した
     /// 再生要求。 `Some(どこから)` の間は最後の load 完了で `fire_pending_play` が再発火する。
     pub pending_play: Option<PlayFrom>,
